+++ projects/core/src/main/kotlin/site/siredvin/turtlematic/computercraft/peripheral/forged/MasonAutomataCorePeripheral.kt
@@ -169,13 +169,13 @@ class MasonAutomataCorePeripheral(turtle: ITurtleAccess, side: TurtleSide, tier:
 
     private fun isEditable(pos: BlockPos): Boolean {
         return peripheralOwner.withPlayer({
-            PeripheraliumPlatform.isBlockProtected(pos, it.level.getBlockState(pos), it)
+            PeripheraliumPlatform.isBlockProtected(pos, it.fakePlayer.level.getBlockState(pos), it.fakePlayer)
         })
     }
 
     private fun findBlock(overwrittenDirection: VerticalDirection?): Pair<Pair<BlockHitResult, BlockState>?, MethodResult?> {
         val hit = peripheralOwner.withPlayer({
-            val hit = FakePlayerProxy(it).findHit(skipEntity = true, skipBlock = false)
+            val hit = it.findHit(skipEntity = true, skipBlock = false)
             if (hit !is BlockHitResult) {
                 return@withPlayer null
             }
@@ -277,7 +277,7 @@ class MasonAutomataCorePeripheral(turtle: ITurtleAccess, side: TurtleSide, tier:
         val level = level!!
         val fakeContainer: Container? = if (mode == TransformInteractionMode.BLOCK) {
             val blockState = peripheralOwner.withPlayer({
-                val hit = FakePlayerProxy(it).findHit(skipEntity = true, skipBlock = false)
+                val hit = it.findHit(skipEntity = true, skipBlock = false)
                 if (hit !is BlockHitResult) {
                     return@withPlayer null
                 }